shellcheck: Make all bash scripts pass shellcheck
This commit adds a "shellcheck" Makefile target, and fixes all of
the bash scripts to pass a shellcheck [1] run. Shellcheck does
static analysis of bash scripts. In the past, it's saved me hours
of debugging things in production, I find it useful to use early
in a projects lifetime. Once we have proper CI/CD for IPDK, we'd
want to run the "shellcheck" target on each pull request before
merging, for example.

[1] https://github.com/koalaman/shellcheck

Signed-off-by: Kyle Mestery <mestery@mestery.com>