ipdk-lib.sh: Remove get_distro_image return value
This removes the return value of get_distro_image(). Using echo was
resulting in failures because it was taking the output of `wget` in
addition to the echo'd return value. It turns out it's not necessary,
because the name of the downloaded image is hard coded anyways.

This allows `ipdk demo` to work perfectly!

Signed-off-by: Kyle Mestery <mestery@mestery.com>