actions: Enable GitHub actions
This enables GitHub Actions to do basic CI on pull requests to the IPDK
repository. For now, it's actually a bit of a brute force approach, in
that we dynamically configure each Runner with the tools we need,
including Docker. We next run shellcheck, followed by the full build.

Some ways to optimize this in the future:

* Build our runner which has all the tools configured and installed.
* Modularize the build a bit to remove redundant steps.
* Work to get patches into upstream projects so we don't have to build
  them each time.

Closes #88

Signed-off-by: Kyle Mestery <mestery@mestery.com>