Add host-target customization capability.
IPDK is planned as target-agnostic set of drivers/tools etc.
Based on that we need to provide an ability to run our services with
different target platforms. host-target container is one of them and
it should be able to detect a disk to run traffic against by means
of SPDK SMA opaque handle. For that purpose, ability to run custom
device exerciser is added.

Signed-off-by: Artsiom Koltun <artsiom.koltun@intel.com>