Remove "python" dependency from creating virtio block device.
create_virtio_blk() used "python" that defaults to Python 2 on Debian,
and that may not be installed on the system. On the other hand,
create_virtio_blk() uses rpc.py, that already depends on
"python3". This patch removes the Python 2 dependency and fixes issues
that arise from using the function in a system without "python"
executable.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>