Change disk_infrastructure functions to python
Previously, all functions used to execute operations described in recipes
or used in tests were written in shell. This approach was error-prone
and more reliable language is used. Rewritten functions were placed in a
new file - disk_infrastructure.py - and formatted with black. An __init__.py
file was created inside of scripts directory to make it a package and easily
import python functions.
Reworked functionality:
- Subsystem creation
- Ramdrive creation and its attachment to subsystem
- Virtio-blk creation
- Virtio-blk deletion
- Checking if a port is open
- Getting/Checking number of Virtio-blk devices
- Checking if a Virtio-blk is attached or not

Signed-off-by: rilnickx <rafalx.ilnicki@intel.com>