Refactor storage host target to cleaner OO-design
IPDK intends to support different target devices. The KVM case exists 
mainly for demonstration purposes. Despite many pros it also has some
disadvantages like lack of VF-support for NVMe.

In this patch the host-target model and code are reworked to provide
a target-agnostic way for PFs and VFs to appear within host OS.
For HW targets, it is required to provide their own SMA handle class
which can supply required information like finding a corresponding
PCI address and information if provided SMA handle refers to a virtual device
or the protocol in use.

PF/VF functionality is covered with unit tests against fake sysfs
with, at this moment, limited functionality.