Run fio against multiple volumes with dedicated jobs.
In the previous version, if a fio is run to exercise all volumes
attached to a device, all of them are exercised within a single job.
However, it produced a joint report which does not show results for
each volume separately. It can be inconvenient for the end user,
if a per-volume statistic is needed.

In this patch, all volumes are exercised within their own fio job,
and a separate report is shown for each device.

Signed-off-by: Artsiom Koltun <artsiom.koltun@intel.com>